+++ packages/object/of.ts
@@ -1,7 +1,6 @@
 import {isFunction, isUsableImmutablePrimitive} from './is';
 import {isset} from './isset';
-import {apply} from '../platform/function';
-import {TypeConstructor} from "../types";
+import {Constructable} from "../types";
 
 /**
  * Creates a value `of` given type;  Checks for one of the following construction strategies (in order listed):
@@ -11,22 +10,17 @@ import {TypeConstructor} from "../types";
  * //      constructor as a function (in cast form;  E.g., `constructor(...args)` )
  * // - Else if constructor is a function, thus far, then calls constructor using
  * //      the `new` keyword (with any passed in args).
-
- * @function module:object.of
- * @param x {*} - Value to derive returned value's type from.
- * @param [args] {...*} - Any args to pass in to matched construction strategy.
- * @returns {*|undefined} - New value of given value's type else `undefined`.
  */
 export const of = <T>(x: T, ...args: any[]): T => {
-    if (!isset(x)) { return undefined; }
-    const constructor: TypeConstructor = x.constructor;
-    if (constructor.of) {
-        return constructor.of(...args);
+    if (!isset(x)) {
+        return undefined;
     }
-    else if (isUsableImmutablePrimitive(x)) {
-        return constructor(args);
-    }
-    else if (isFunction(constructor)) {
+    const constructor = x.constructor as Constructable;
+    if (constructor['of']) {
+        return constructor['of'](...args);
+    } else if (isUsableImmutablePrimitive(x)) {
+        return (constructor as unknown as Function)(args);
+    } else if (isFunction(constructor)) {
         return new constructor(...args);
     }
     return undefined;