+++ packages/platform/function/types.ts
@@ -1,5 +1,5 @@
 import {CurryOf2} from "../../function";
 
-export type ApplyFunc = CurryOf2<Function, any[], any>;
+export type ApplyFunc = CurryOf2<Function, unknown[], unknown>;
 
-export type CallFunc = CurryOf2<Function, any, any>;
+export type CallFunc = CurryOf2<Function, unknown, unknown>;